feat: track multiple vehicles for the same TripUpdate (#90)
The vehicle itself is only on a single trip at a time, but a TripUpdate can
predict that a vehicle will be running that trip in the future. Previously,
we'd throw that information away, relying on the information from the
VehiclePosition to hold the trip information. Now, we also keep track of
vehicle's ID for the TripUpdate, and use that if there isn't a
VehiclePosition with that trip ID.